requirements.txt
autobahn
dbus-python
qrcode[pil]

